# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ESG ProofOps API
  version: 1.0.0
  description: Implementation contract, not a deployed service. Domain v2.0 is authoritative.
servers:
- url: /
paths:
...
  /v1/runs/{run_id}/exports:
    post:
      operationId: export_create
      summary: export_create
      x-minimum-role: viewer
      x-rate-limit: 10/min/user
      x-idempotency-required: true
      security:
      - sessionCookie: []
      responses:
        '202':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Export'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden or CSRF failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found or tenant mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict / gate blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          description: Stale revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Schema or domain validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate/budget limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 16
          maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportCreate'
  /v1/exports/{export_id}:
    get:
      operationId: export_get
      summary: export_get
      x-minimum-role: viewer
      x-rate-limit: 120/min/user
      x-idempotency-required: false
      security:
      - sessionCookie: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Export'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden or CSRF failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found or tenant mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict / gate blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          description: Stale revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Schema or domain validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate/budget limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: export_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /v1/exports/{export_id}/download:
    post:
      operationId: export_download
      summary: export_download
      x-minimum-role: viewer
      x-rate-limit: 60/min/user
      x-idempotency-required: false
      security:
      - sessionCookie: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Download'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden or CSRF failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found or tenant mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict / gate blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          description: Stale revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Schema or domain validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate/budget limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: export_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-proofops_session
  schemas:
    Error:
      type: object
      additionalProperties: false
      properties:
        error:
          type: object
          additionalProperties: false
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
              format: uuid
            retryable: &id003
              type: boolean
            details:
              type: object
              additionalProperties: false
              properties:
                field:
                  type: string
                reason:
                  type: string
                current_revision:
                  type: integer
                  minimum: 1
                retry_after_seconds:
                  type: integer
                  minimum: 0
          required:
          - code
          - message
          - request_id
          - retryable
      required:
      - error
...
    ExportCreate:
      type: object
      additionalProperties: false
      properties:
        formats:
          type: array
          items:
            type: string
            enum:
            - json
            - csv
            - html
          minItems: 1
          uniqueItems: true
        allow_partial: *id003
      required:
      - formats
      - allow_partial
    Export:
      type: object
      additionalProperties: false
      properties:
        export_id:
          type: string
          format: uuid
        run_id:
          type: string
          format: uuid
        state:
          type: string
          enum:
          - queued
          - building
          - ready
          - failed
        snapshot_epoch:
          type: integer
          minimum: 0
        partial: *id003
        manifest_sha256:
          anyOf:
          - *id002
          - type: 'null'
        created_at: *id001
      required:
      - export_id
      - run_id
      - state
      - snapshot_epoch
      - partial
      - manifest_sha256
      - created_at
    Download:
      type: object
      additionalProperties: false
      properties:
        url:
          type: string
        expires_at: *id001
        sha256: *id002
      required:
      - url
      - expires_at
      - sha256
...